pyproject.toml
[project]
name = "criatons"
version = "1.0.0"
description = "Agente autônomo de vendas via WhatsApp para MedReview"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.50.0",
    "flask>=3.0.0",
    "gunicorn>=21.0.0",
    "httpx>=0.24.0",
    "supabase>=2.3.0",
    "gevent>=23.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
